Move IsHit extension method to Extensions namespace
