+++ HomeRunTracker.Backend/Services/PitcherGameScoreService.cs
@@ -1,4 +1,5 @@
-﻿using HomeRunTracker.Common.Enums;
+﻿using HomeRunTracker.Backend.Extensions;
+using HomeRunTracker.Common.Enums;
 using HomeRunTracker.Common.Models.Details;
 using HomeRunTracker.Common.Models.Internal;
