+++ HomeRunTracker.Common/Enums/EPlayResult.cs
@@ -13,18 +13,3 @@ public enum EPlayResult
     Triple,
     HomeRun,
 }
-
-public static class EPlayResultExtensions
-{
-    public static bool IsHit(this EPlayResult playResult)
-    {
-        return playResult switch
-        {
-            EPlayResult.Single => true,
-            EPlayResult.Double => true,
-            EPlayResult.Triple => true,
-            EPlayResult.HomeRun => true,
-            _ => false
-        };
-    }
-}
